refactor(keyboard): ショートカットregistryをlib/keyboardへ移す
registry と key-label は logger 以外に依存が無く、実際の global keydown
listener も GlobalOverlays から app 全体へ 1 回だけ登録されている。名前と
置き場だけが calendar を指していた。

ShortcutHelpLabelKey の閉じた union（全て calendar.shortcuts.actions.*）は
lib に calendar の語彙を持ち込むため string へ広げる。宣言の一意性は後続の
カタログが担保する。ViewSwitcher の deep relative import も alias へ揃えた。
