+++ apps/product/src/lib/keyboard/shortcut-key-label.ts
